Added getStudentsByDate()
Currently, it just sorts by any date before the one entered.
Things to keep in mind: Make sure date format in string is mm/dd/yyyy
and not mm/d/yyyy etc.
And there may be a problem with leading zeros when trying to convert
string to int. Need to explore this in more detail and probably use a
date data type.